+++ Source/Internal/Base.h
