+++ Source/Naive/astar.cpp
@@ -1,11 +1,11 @@
 // Source Code: https://github.com/hit9/quadtree-pathfinding
 // License: BSD. Chao Wang, hit9[At]icloud.com.
 
-#include "astar.hpp"
+#include "Astar.h"
 
 #include <cassert>
 
-#include "../Internal/pathfinder_astar.hpp"
+#include "../Internal/PathfinderAstar.h"
 
 namespace qdpf
 {
