+++ Source/Naive/flowfield.cpp
@@ -1,12 +1,12 @@
 // Source Code: https://github.com/hit9/quadtree-pathfinding
 // License: BSD. Chao Wang, hit9[At]icloud.com.
 
-#include "flowfield.hpp"
+#include "Flowfield.h"
 
 #include <cassert>
 
-#include "../Internal/base.hpp"
-#include "../Internal/pathfinder_flowfield.hpp"
+#include "../Internal/Base.h"
+#include "../Internal/PathfinderFlowfield.h"
 
 namespace qdpf
 {
